Realised DI extentions for Application and Infrastructure.Blockchain

+++ Voting.Application/Events/CandidateAddedEventArgs.cs
@@ -3,10 +3,10 @@
     /// <summary>
     /// Аргументы события добавления кандидата.
     /// </summary>
-    public sealed class CandidateAddedEventArgs(ulong sessionId, ulong candidateId, string name) : EventArgs
+    public sealed class CandidateAddedEventArgs(uint sessionId, uint candidateId, string name) : EventArgs
     {
-        public ulong SessionId { get; } = sessionId;
-        public ulong CandidateId { get; } = candidateId;
+        public uint SessionId { get; } = sessionId;
+        public uint CandidateId { get; } = candidateId;
         public string Name { get; } = name;
     }
 }
