+++ Voting.Application/Events/VotingEndedEventArgs.cs
@@ -4,9 +4,9 @@ namespace Voting.Application.Events
     /// <summary>
     /// Аргументы события окончания голосования.
     /// </summary>
-    public sealed class VotingEndedEventArgs(ulong sessionId, DateTime endTimeUtc) : EventArgs
+    public sealed class VotingEndedEventArgs(uint sessionId, DateTime endTimeUtc) : EventArgs
     {
-        public ulong SessionId { get; } = sessionId;
+        public uint SessionId { get; } = sessionId;
         public DateTime EndTimeUtc { get; } = endTimeUtc;
     }
 }
